+++ [T]LogicLab/TB_TinyDancer/Button/ButtonTB.h
@@ -5,6 +5,6 @@
 #define MAX_BUTTON_VALUE  200
 
 X_Boolean CallFunction(uint8_t func_num,X_Void * p_param);
-uint32_t  GetCurrentButtonValue(X_Void);
+uint16_t  GetCurrentButtonValue(X_Void);
 
 #endif
